third_party/WebKit/LayoutTests/dom/domparsing/dom-parse-serialize-display-expected.txt

DOMParser/XMLSerializer test
The "text to parse" and "document object serialized" boxes should show the same text, and it should be an XML document, not "@@No result@@".
text to parse
<?xml version="1.0"?>
<?xml-stylesheet href="display.css" type="text/css"?>
<!DOCTYPE doc [
<!ATTLIST d id ID #IMPLIED>
]>
<doc>
  <foo xmlns="foobar">One</foo> <x:bar xmlns:x="barfoo">Two</x:bar>
  <d id="id3">Three</d>
</doc>



document object
[object XMLDocument]


document object serialized
<?xml version="1.0"?><?xml-stylesheet href="display.css" type="text/css"?><!DOCTYPE doc><doc>
  <foo xmlns="foobar">One</foo> <x:bar xmlns:x="barfoo">Two</x:bar>
  <d id="id3">Three</d>
</doc>
